requirements.txt
pandas
palmerpenguins
streamlit 
numpy 
scikit-learn 
plotly 
pillow